WIP SimpleRDF->grapoi
